Fixes to ejabberd recipes and courses/views
* Rewrite of ejabberd installation recipes to avoid wiping mnesia tables
  unless this is a first-time install.
* Also, recipe uses octal mode notation only (strings removed)
* Also, template for ejabberd defaults file is actually a template
* Fix in courses/views to use hasattr instead of getattr

+++ chef/cookbooks/class2go-ejabberd/recipes/default.rb
@@ -1,20 +1,15 @@
-# Installing ejabberd and configuring it is fiddly. A lot of material gets 
-# created when it's nonexistent, which gives things a slightly magical feeling,
-# and not in a good way. Its especially problematic if you start ejabberd with
-# one set of settings (such as hostname), and then restart it later with a
-# different set. The following set of steps are in the following order because
-# this seems to satisfy the requirements for having working ejabberd most
-# easily.
-#
-# This set of recipes must be run after base-ubuntu, so we have a hostname.
+# Default case is just upgrading the package or changing configuration.
+# Note that if you case the host's Name tag to change, things will break.
+# The easiest fix in this case is to uninstall ejabberd, remove 
+# /var/lib/ejabberd, and then re-run chef-client.
+WIPE_MNESIA_TABLES = false
 
-# Remove ejabberd and completely destroy existing configuration 
-# before doing other things
-service "ejabberd" do
-    action :stop
-end
-package "ejabberd" do
-    action :purge
+if not FileTest.directory?('/var/lib/ejabberd')
+    # When the ejabberd package gets installed, ejabberd gets started, and a
+    # bunch of mnesia tables get created with the wrong data in them. So if 
+    # we've never had ejabberd on this box, we should remember to wipe this 
+    # directory during our reconfiguration stage (below).
+    WIPE_MNESIA_TABLES = true
 end
 
 # Now install ejabberd and dependencies (Big!) with debian-default config
@@ -22,11 +17,12 @@ package "ejabberd" do
     action :install
 end
 
-# ejabberd is started at installation; stop it and wipe runtime files so we
-# can reconfigure
+# ejabberd is started at installation. Let's stop it while we reconfigure
 service "ejabberd" do
     action :stop
 end
+# Next two stanzas are because ubuntu's service scripts are too dumb to kill
+# orphaned erlang processes after their parents crash
 execute "/usr/bin/killall beam" do
     user "root"
     action :run
@@ -37,10 +33,12 @@ execute "/usr/bin/killall epmd" do
     action :run
     returns [0, 1]
 end
-execute "/bin/rm -rf /var/lib/ejabberd/*" do
-    user "root"
-    action :run
-    returns [0, 1]
+if WIPE_MNESIA_TABLES
+    execute "/bin/rm -rf /var/lib/ejabberd/*" do
+        user "root"
+        action :run
+        returns [0, 1]
+    end
 end
 
 # Configure ejabberd with our preferred settings
@@ -54,7 +52,7 @@ end
 directory "/etc/default" do
     owner 'root'
     group 'root'
-    mode '0755'
+    mode 00755
     action :create
 end
 
@@ -62,7 +60,7 @@ template "/etc/default/ejabberd" do
     source "etc-default-ejabberd.erb"
     owner 'root'
     group 'root'
-    mode '0644'
+    mode 00644
 end
 
 template "/etc/ejabberd/ejabberd.cfg" do
@@ -76,10 +74,10 @@ cookbook_file "/etc/ejabberd/inetrc" do
     source "etc-ejabberd-inetrc"
     owner "ejabberd"
     group "ejabberd"
-    mode "0644"
+    mode 00644
 end
 
-# Ok, now it should start ok. Run the ejabberd server
+# Now it's safe to start the server for real
 service "ejabberd" do
     action :start
 end
